TAG-70: Del opp en avtaleside i Kontaktinformasjon og Avtale
