+++ src/AvtaleSide/AvtaleModell.ts
