+++ src/AvtaleSide/AvtaleOversikt.tsx
@@ -1,5 +1,5 @@
 import * as moment from 'moment';
-import KnappBase from 'nav-frontend-knapper';
+import { Knapp } from 'nav-frontend-knapper';
 import PanelBase from 'nav-frontend-paneler';
 import * as React from 'react';
 import { Link } from 'react-router-dom';
@@ -10,7 +10,7 @@ import AvtaleModell, { tomAvtale } from './AvtaleModell';
 const Avtaler = (props: { avtaler: AvtaleModell[] }) => {
     const avtaleLinker = props.avtaler.map((avtale: AvtaleModell) => (
         <li key={avtale.id}>
-            <Link to={'/avtale/' + avtale.id}>
+            <Link to={'/avtale/' + avtale.id + '/kontaktinformasjon'}>
                 Avtale (opprettet: {avtale.opprettetTidspunkt})
             </Link>
         </li>
@@ -27,7 +27,7 @@ class AvtaleOversikt extends React.Component<
         avtaler: [],
     };
 
-    componentWillMount() {
+    componentDidMount() {
         this.hentAvtaler();
     }
 
@@ -44,7 +44,9 @@ class AvtaleOversikt extends React.Component<
             opprettetTidspunkt: moment().format('DD.MM.YYYY HH:mm:ss'),
         };
         avtaleRef.set(avtale).then(() => {
-            this.props.history.push('/avtale/' + avtaleId);
+            this.props.history.push(
+                '/avtale/' + avtaleId + '/kontaktinformasjon'
+            );
         });
     };
 
@@ -71,13 +73,9 @@ class AvtaleOversikt extends React.Component<
         return (
             <PanelBase>
                 <Avtaler avtaler={this.state.avtaler} />
-                <KnappBase
-                    type="standard"
-                    disabled={false}
-                    onClick={this.opprettAvtale}
-                >
+                <Knapp disabled={false} onClick={this.opprettAvtale}>
                     Opprett avtale
-                </KnappBase>
+                </Knapp>
             </PanelBase>
         );
     }
