+++ src/AvtaleSide/AvtaleStegProps.ts
@@ -4,5 +4,5 @@ import AvtaleModell from './AvtaleModell';
 export default interface AvtaleStegProps {
     label: string;
     form: AvtaleModell;
-    handleChange: EventHandler<any>;
+    oppdaterAvtale: EventHandler<any>;
 }
