+++ src/index.tsx
@@ -1,8 +1,14 @@
 import * as React from 'react';
 import * as ReactDOM from 'react-dom';
 import App from './App';
+import { Switch } from 'react-router';
+import { BrowserRouter } from 'react-router-dom';
 
 ReactDOM.render(
-  <App />,
-  document.getElementById('root') as HTMLElement
+    <BrowserRouter>
+        <Switch>
+            <App />
+        </Switch>
+    </BrowserRouter>,
+    document.getElementById('root') as HTMLElement
 );
